Provide Docker image for building pycdio wheels
Add a Dockerfile for building a container image which can be used for
building pycdio manylinux2014_x86_64 wheels (compliant with PEP 599).
This includes a build script, default settings for pip invoked inside
the container, and a README with usage instructions.

Ignore directory where the wheel files are stored and update Makefile
to remove it in the "clean" target.